requirements.txt
django==2.0.5
python-twitter==3.4.1